cross gen is running with the benchmark mode
